Create Demonstrates the use of data members and member functions
1.Data Members:

In the person class, you have two data members:
char name[MAX_CHAR]: This is a character array to store the name of the person. It represents a data member for storing the name.

int age: This is an integer variable to store the age of the person. It represents another data member for storing the age.

2.Member Functions:

void get(char n[], int a): This member function takes two parameters - a character array representing the name and an integer representing the age. It copies these values into the corresponding data members of the person object.

void put(): This member function doesn't take any parameters but displays the values of the data members (name and age) of the person object.

3.Main Function:

In the main function, you create an instance of the person class named PER.
You then call the get member function to set the name and age of the PER object to "Shovin" and 18, respectively.
Finally, you call the put member function to display the name and age of the PER object.